api/requirements.txt
Flask             == 0.10.1
Flask-Injector    == 0.4.0
Flask-SQLAlchemy  == 0.16
Flask-Script      == 2.0.5
injector          == 0.9.0
psycopg2          == 2.5.4
schema            == 0.3.1

Flask-Testing     == 0.4.2
nose              == 1.3.4
